refactor(builders): put MySQL on the shared build path
MySQL carried its own copy of the whole build. Only its mutation half is
genuinely different — its writes return no rows, so every mutation reports
`{ isSuccess }` instead of the rows it touched. Everything else was a copy:
table discovery, the relation graph, the per-build registries and type cache,
the filter context, each table's types, the read fields, and the relation
resolvers at the end.

The copy is what made #90 ship half-delivered: MySQL accepted
`features.uniqueKeyFilters` and generated nothing, because the unique-key map
was built in the other builder. Enabling it on MySQL now produces the key
field, with a regression test that says so.

The three dialect-independent stretches move to `build-context.ts` —
`prepareBuild` (the prelude), `addReadFields` (both selects, aggregate,
groupBy), `finalizeBuild` (the mutation roster and relation resolvers) —
behind a `SchemaBuildAdapter` whose `returnsRows` flag decides whether a
table's types carry the `${Table}Item` outputs. `addReadFields` returns the
types it built rather than registering them, so each caller keeps its own
type-map order and the printed SDL does not drift.

Net: -743 lines across the three builders.
